requirements.txt
alabaster==0.7.12
autopep8==1.5
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
docutils==0.16
idna==2.9
imagesize==1.2.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.3
pycodestyle==2.5.0
pycrypto==2.6.1
Pygments==2.6.1
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.8
